+++ docs/source/tutorial.md
@@ -1,11 +1,11 @@
 # Tutorial
 
-This page covers brief tutorials and notes on how to use pycytominer-transform.
+This page covers brief tutorials and notes on how to use CytoTable.
 
 ## CellProfiler CSV Output to Parquet
 
 [CellProfiler](https://cellprofiler.org/) pipelines or projects may produce various CSV-based compartment output (for example, "Cells.csv", "Cytoplasm.csv", etc.).
-Pycytominer-transform converts this data to Parquet from local or object-storage based locations.
+CytoTable converts this data to Parquet from local or object-storage based locations.
 
 Files with similar names nested within sub-folders will be concatenated by default (appended to the end of each data file) together and used to create a single Parquet file per compartment.
 For example: if we have `folder/subfolder_a/cells.csv` and `folder/subfolder_b/cells.csv`, using `convert(source_path="folder", ...)` will result in `folder.cells.parquet` (unless `concat=False`).
@@ -13,7 +13,7 @@ For example: if we have `folder/subfolder_a/cells.csv` and `folder/subfolder_b/c
 For example, see below:
 
 ```python
-from pycytominer_transform import convert
+from cytotable import convert
 
 # using a local path with cellprofiler csv presets
 convert(
