+++ readme.md
@@ -1,20 +1,20 @@
-# pycytominer-transform
+# CytoTable
 
 ![dataflow](docs/source/_static/dataflow.svg)
 _Diagram showing data flow relative to this project._
 
 ## Summary
 
-pycytominer-transform enables single-cell morphology data analysis by cleaning and transforming CellProfiler (`.csv` or `.sqlite`), cytominer-database (`.sqlite`), and DeepProfiler (`.npz`) output data at scale.
-pycytominer-transform creates parquet files for both independent analysis and for input into [Pycytominer](https://github.com/cytomining/pycytominer).
+CytoTable enables single-cell morphology data analysis by cleaning and transforming CellProfiler (`.csv` or `.sqlite`), cytominer-database (`.sqlite`), and DeepProfiler (`.npz`) output data at scale.
+CytoTable creates parquet files for both independent analysis and for input into [Pycytominer](https://github.com/cytomining/pycytominer).
 The Parquet files will have a unified and documented data model, including referenceable schema where appropriate (for validation within Pycytominer or other projects).
 
 ## Installation
 
-Install pycytominer-transform with the following command:
+Install CytoTable with the following command:
 
 ```shell
-pip install git+https://github.com/cytomining/pycytominer-transform.git
+pip install git+https://github.com/cytomining/CytoTable.git
 ```
 
 ## Contributing, Development, and Testing
